+++ lib/wavefront-sdk/validators.rb
@@ -568,6 +568,32 @@ module Wavefront
 
       raise Wavefront::Exception::InvalidPermission, id, id
     end
+
+    # Ensure the given argument is a valid ingestion policy ID
+    # @param id [String]
+    # @raise Wavefront::Exception::InvalidIngestionPolicyId if the
+    #   ID is not valid
+    #
+    def wf_ingestionpolicy_id?(id)
+      return true if id.is_a?(String) && id =~ /^[a-z0-9\-_]+-\d{13}$/
+
+      raise Wavefront::Exception::InvalidIngestionPolicyId, id
+    end
+
+    # Ensure the given argument is a valid User or SystemAccount ID.
+    # @param id [String]
+    # @raise Wavefront::Exception::InvalidAccountId if the
+    #   ID is not valid
+    #
+    def wf_account_id?(id)
+      return true if wf_user_id?(id)
+    rescue Wavefront::Exception::InvalidUserId
+      begin
+        return true if wf_serviceaccount_id?(id)
+      rescue Wavefront::Exception::InvalidServiceAccountId
+        raise Wavefront::Exception::InvalidAccountId, id
+      end
+    end
   end
   # rubocop:enable Metrics/ModuleLength
 end