+++ HISTORY.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## 3.7.0 (2020-01-21)
+## 3.7.0 (2020-01-23)
+* Add `Account`, `Usage` and `IngestionPolicy` classes.
+* Allow modification of `Wavefront::Response`'s `response` object.
 * Add `User#validate_users` method.
 
 ## 3.6.1 (2020-01-15)